+++ packages/at_chops/lib/src/key/impl/at_pkam_key_pair.dart
@@ -1,5 +1,8 @@
 import 'package:at_chops/src/key/keys.dart';
 
+@Deprecated(
+    'Use RsaKeyPair for RSA key material instead. This compatibility wrapper '
+    'will be removed in the next major release.')
 class AtPkamKeyPair extends AsymmetricKeyPair {
   AtPkamKeyPair.create(super.publicKey, super.privateKey) : super.create();
 }